data/app/pyproject.toml
[tool.poetry]
name = "cache_model_project"
version = "0.1.0"
description = ""
authors = ["zed"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.0.post1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
aiohttp = "^3.9.3"
pydantic-settings = "^2.1.0"
pycryptodome = "^3.20.0"
alembic = "^1.13.1"
python-ulid = "^2.2.0"
sqladmin = "^0.16.1"
fastapi-cache2 = "^0.2.1"
redis = "^5.0.3"
aioredis = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"

[tool.ruff]
line-length = 120
target-version = "py310"
fix = false
show-fixes = true
output-format = "grouped"
extend-exclude = ["*.pyi"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
  "A",      # flake8-builtins
  "ANN",    # flake8-annotations
  "ARG",    # flake8-unused-arguments
  "ASYNC",  # flake8-async
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "C90",    # mccabe
  "DTZ",    # flake8-datetimez
  "E",      # pycodestyl
  "EM",     # flake8-errmsg
  "ERA",    # eradicate
  "F",      # pyflakes
  "FBT",    # flake8-boolean-trap
  "I",      # isort
  "INT",    # flake8-gettext
  "N",      # pep8-naming
  "PL",     # Pylint
  "PERF",   # Perflint
  "PT",     # flake8-pytest-style
  "RET",    # flake8-return
  "RUF",    # Ruff-specific rules
  "S",      # flake8-bandit
  "SLF",    # flake8-flake8-self
  "T20",    # flake8-print
  "TRY",    # tryceratops
  "SIM",    # flake8-simplify
  "UP",     # pyupgrade
  "W",      # pycodestyle
]
ignore = ["ANN101"]
ignore-init-module-imports = true

[tool.ruff.lint.extend-per-file-ignores]
"**/{tests,docs,tools}/*" = ["E402", "S101", "ANN401", "ANN001", "S603", "ARG001"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
lines-between-types = 1
lines-after-imports = 2

[tool.ruff.lint.mccabe]
# Flag errors (C901) whenever the complexity level exceeds 6.
max-complexity = 6

[tool.ruff.lint.pylint]
allow-dunder-method-names = ["__tablename __", "__ table_args__"]

[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"